Adding Visual studio compile for the rest of the projet

+++ sources/softs/generate_data/main.cpp
@@ -72,7 +72,7 @@ int main(int argc, char** argv)
 		f << "#PARAM_IN COS_TH" << std::endl;
 		for(int i=0; i<nbx; ++i)
 		{
-				const float    x = i / (float)nbx ;
+                const double   x = i / (float)nbx ;
 //				const float d[3] = {0.1, 0.0, 0.5};
 				const float d[3] = {0.0, 0.0, 0.0};
 				const float z1 = d[0] + 0.2 * std::pow(x, 1.5) ;
@@ -91,8 +91,8 @@ int main(int argc, char** argv)
 		f << "#PARAM_IN COS_TH" << std::endl;
 		for(int i=0; i<nbx; ++i)
 		{
-				const float x = i / (float)nbx ;
-				const float z = 0.1 + 0.5 * std::pow(x, 1.5) ;
+                const double x = i / (float)nbx ;
+                const double z = 0.1 + 0.5 * std::pow(x, 1.5) ;
 			
 				f << x << "\t" << z << std::endl ;
 		}
